backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
numpy==2.4.3
pandas==3.0.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
model-bakery==1.17.0
